Extract ZoomableImageView to separate file for SwiftLint file_length
